requirements.txt
django-widget-tweaks==1.4.1
django_social_share
